requirements.txt
pandas>=1.3.0
numpy>=1.20.0
pysam>=0.16.0
biopython>=1.79
click>=8.0.0